Remove v1 from packages
