+++ service/controller/chart/resource/release/update.go
@@ -8,13 +8,14 @@ import (
 	"github.com/giantswarm/helmclient"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/operatorkit/controller/context/resourcecanceledcontext"
+	"github.com/giantswarm/operatorkit/resource/crud"
 	"k8s.io/helm/pkg/helm"
 
-	"github.com/giantswarm/chart-operator/service/controller/chart/v1/controllercontext"
-	"github.com/giantswarm/chart-operator/service/controller/chart/v1/key"
+	"github.com/giantswarm/chart-operator/service/controller/chart/controllercontext"
+	"github.com/giantswarm/chart-operator/service/controller/chart/key"
 )
 
-func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange interface{}) error {
+func (r *Resource) ApplyUpdateChange(ctx context.Context, obj, updateChange interface{}) error {
 	cr, err := key.ToCustomResource(obj)
 	if err != nil {
 		return microerror.Mask(err)
@@ -24,17 +25,17 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 		return microerror.Mask(err)
 	}
 
-	releaseState, err := toReleaseState(createChange)
+	releaseState, err := toReleaseState(updateChange)
 	if err != nil {
 		return microerror.Mask(err)
 	}
 
+	upgradeForce := key.HasForceUpgradeAnnotation(cr)
+
 	if releaseState.Name != "" {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("creating release %#q", releaseState.Name))
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("updating release %#q with force == %t", releaseState.Name, upgradeForce))
 
-		ns := key.Namespace(cr)
 		tarballURL := key.TarballURL(cr)
-
 		tarballPath, err := r.helmClient.PullChartTarball(ctx, tarballURL)
 		if helmclient.IsPullChartFailedError(err) {
 			reason := fmt.Sprintf("pulling chart %#q failed", tarballURL)
@@ -73,16 +74,18 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 
 		ch := make(chan error)
 
-		// We create the helm release but with a short timeout so we don't
+		// We update the helm release but with a short timeout so we don't
 		// block reconciling other CRs. This gives time to make the port
 		// forwarding connection to the Tiller API.
 		//
-		// If we do timeout the install will continue in the background.
+		// If we do timeout the update will continue in the background.
 		// We will check the progress in the next reconciliation loop.
 		go func() {
-			// We need to pass the ValueOverrides option to make the install process
+			// We need to pass the ValueOverrides option to make the update process
 			// use the default values and prevent errors on nested values.
-			err = r.helmClient.InstallReleaseFromTarball(ctx, tarballPath, ns, helm.ReleaseName(releaseState.Name), helm.ValueOverrides(releaseState.ValuesYAML))
+			err = r.helmClient.UpdateReleaseFromTarball(ctx, releaseState.Name, tarballPath,
+				helm.UpdateValueOverrides(releaseState.ValuesYAML),
+				helm.UpgradeForce(upgradeForce))
 			close(ch)
 		}()
 
@@ -97,30 +100,30 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 				return microerror.Mask(err)
 			}
 
-			r.logger.LogCtx(ctx, "level", "debug", "message", "release still being created")
+			r.logger.LogCtx(ctx, "level", "debug", "message", "release still being updated")
 			r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 			return nil
 		}
 
 		if err != nil {
-			reason := err.Error()
 			r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("helm release %#q failed", releaseState.Name), "stack", microerror.JSON(err))
 
 			releaseContent, err := r.helmClient.GetReleaseContent(ctx, releaseState.Name)
 			if helmclient.IsReleaseNotFound(err) {
-				reason = fmt.Sprintf("helm error: (%s)", reason)
+				reason := fmt.Sprintf("release %#q not found", releaseState.Name)
 				addStatusToContext(cc, reason, releaseNotInstalledStatus)
 
+				r.logger.LogCtx(ctx, "level", "warning", "message", reason, "stack", microerror.JSON(err))
 				r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 				resourcecanceledcontext.SetCanceled(ctx)
 				return nil
+
 			} else if err != nil {
 				return microerror.Mask(err)
 			}
-
 			// Release is failed so the status resource will check the Helm release.
 			if releaseContent.Status == helmFailedStatus {
-				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("failed to create release %#q", releaseContent.Name))
+				r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("failed to update release %#q", releaseContent.Name))
 				r.logger.LogCtx(ctx, "level", "debug", "message", "canceling resource")
 				resourcecanceledcontext.SetCanceled(ctx)
 				return nil
@@ -133,33 +136,53 @@ func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange inte
 			return microerror.Mask(err)
 		}
 
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("created release %#q", releaseState.Name))
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("updated release %#q", releaseState.Name))
 	}
 
 	return nil
 }
 
-func (r *Resource) newCreateChange(ctx context.Context, obj, currentState, desiredState interface{}) (interface{}, error) {
+func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*crud.Patch, error) {
+	create, err := r.newCreateChange(ctx, obj, currentState, desiredState)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+	update, err := r.newUpdateChange(ctx, obj, currentState, desiredState)
+	if err != nil {
+		return nil, microerror.Mask(err)
+	}
+
+	patch := crud.NewPatch()
+	patch.SetCreateChange(create)
+	patch.SetUpdateChange(update)
+
+	return patch, nil
+}
+
+func (r *Resource) newUpdateChange(ctx context.Context, obj, currentState, desiredState interface{}) (interface{}, error) {
 	currentReleaseState, err := toReleaseState(currentState)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
+
 	desiredReleaseState, err := toReleaseState(desiredState)
 	if err != nil {
 		return nil, microerror.Mask(err)
 	}
 
-	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("finding out if the %#q release has to be created", desiredReleaseState.Name))
+	r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("finding out if the %#q release has to be updated", desiredReleaseState.Name))
 
-	createState := &ReleaseState{}
-
-	if isEmpty(currentReleaseState) {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("the %#q release needs to be created", desiredReleaseState.Name))
-
-		createState = &desiredReleaseState
-	} else {
-		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("the %#q release does not need to be created", desiredReleaseState.Name))
+	if isReleaseInTransitionState(currentReleaseState) {
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("the %#q release is in status %#q and cannot be updated", desiredReleaseState.Name, currentReleaseState.Status))
+		return nil, nil
 	}
 
-	return createState, nil
+	if isReleaseModified(currentReleaseState, desiredReleaseState) {
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("the %#q release has to be updated", desiredReleaseState.Name))
+		return &desiredReleaseState, nil
+	} else {
+		r.logger.LogCtx(ctx, "level", "debug", "message", fmt.Sprintf("the %#q release does not have to be updated", desiredReleaseState.Name))
+	}
+
+	return nil, nil
 }
